Refactor main.rb
Add storage.rb to handle all storing methods
Add app.rb to handle all app functionalities
